refactor: 테스트 케이스 추가
